snitz 2019 revision 1
I changed the director name to "snitz_2019" to be consistent with your other directory naming convention.

The manuscript states that the same descriptors were given to both SmellSpace participants and control group, so it appears that non-uniformity in published .csv files was probably a mistake. I've used the terms in SmellSpace data for both behavior csv files.